Rename crashes_full.rs back to crashes.rs
Per review feedback, the original name is clearer since it's the
default (non-FreeBSD) implementation.
